refactor(Montant): déplace les opérations et comparaisons dans MontantRécurrent et MontantPonctuel
