+++ site/source/contextes/économie-collaborative/domaine/location-de-meublé/revenu-net.test.ts
@@ -1,7 +1,8 @@
 import { Either, Equal, pipe } from 'effect'
 import { describe, expect, it } from 'vitest'
 
-import { eurosParAn, moins } from '@/domaine/Montant'
+import { eurosParAn } from '@/domaine/Montant'
+import { moins } from '@/domaine/MontantRécurrent'
 
 import { calculeCotisations } from './cotisations'
 import { calculeRevenuNet } from './revenu-net'
